update sql order backdate

+++ database/migrations/SQL/order_backdate.sql
@@ -5,7 +5,7 @@ ADD COLUMN `is_backdate` INT DEFAULT 0;
 -- For the 'order_items' table
 ALTER TABLE `order_items`
 ADD COLUMN `is_backdate` INT DEFAULT 0,
-ADD COLUMN `product_id` INT UNSIGNED,
+ADD COLUMN `product_id` BIGINT UNSIGNED,
 ADD INDEX (`product_id`);
 
 ALTER TABLE `order_items`
